refactoring
rename components
